Updated the faq module

+++ default_www/backend/modules/faq/actions/delete_category.php
@@ -7,6 +7,7 @@
  * @subpackage	faq
  *
  * @author		Lester Lievens <lester@netlash.com>
+ * @author		Annelies Van Extergem <annelies@netlash.com>
  * @since		2.1
  */
 class BackendFaqDeleteCategory extends BackendBaseActionDelete
@@ -24,20 +25,24 @@ class BackendFaqDeleteCategory extends BackendBaseActionDelete
 		// does the item exist
 		if($this->id !== null && BackendFaqModel::existsCategory($this->id))
 		{
-			// call parent, this will probably add some general CSS/JS or other required files
-			parent::execute();
+			// get data
+			$this->record = (array) BackendFaqModel::getCategory($this->id);
 
-			// get item
-			$this->record = BackendFaqModel::getCategory($this->id);
+			// allowed to delete the category?
+			if(BackendFaqModel::deleteCategoryAllowed($this->id))
+			{
+				// call parent, this will probably add some general CSS/JS or other required files
+				parent::execute();
 
-			// delete item
-			BackendFaqModel::deleteCategory($this->id);
+				// delete item
+				BackendFaqModel::deleteCategory($this->id);
 
-			// trigger event
-			BackendModel::triggerEvent($this->getModule(), 'after_delete_category', array('id' => $this->id));
+				// category was deleted, so redirect
+				$this->redirect(BackendModel::createURLForAction('categories') . '&report=deleted-category&var=' . urlencode($this->record['title']));
+			}
 
-			// item was deleted, so redirect
-			$this->redirect(BackendModel::createURLForAction('categories') . '&report=deleted&var=' . urlencode($this->record['name']));
+			// delete category not allowed
+			else $this->redirect(BackendModel::createURLForAction('categories') . '&error=delete-category-not-allowed&var=' . urlencode($this->record['title']));
 		}
 
 		// something went wrong
